requirements.txt
fastmcp>=2.0.0
ag2[mcp]>=0.4.0r
beeai-framework>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
rich>=13.0.0
